saas-backend/requirements.txt
# FastAPI + Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# Utils
python-dotenv>=1.0.1

# Dev
httpx>=0.28.0
pytest>=8.3.0
